refactoring by claude
