+++ CLAUDE.md
@@ -7,6 +7,7 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 - DO NOT edit .gitignore
 - DO NOT edit the contents of the .git directory
 - DO NOT edit generated files
+- Prefer functional style without side effect.
 
 ## Project Overview
